Test Mount & devtmpfs
This patch includes the package "mount" in the CI tests to ensure that
no errors are introduced.

This patch also fixes the issue raised by @codenrhoden in #20 regarding
device mounts and devtmpfs
