+++ mount/mount_test.go
@@ -1,9 +1,73 @@
 package mount
 
 import (
+	"io/ioutil"
+	"os"
 	"strings"
 	"testing"
 )
+
+func TestBindMount(t *testing.T) {
+	src, err := ioutil.TempDir("", "")
+	if err != nil {
+		t.Fatal(err)
+	}
+	tgt, err := ioutil.TempDir("", "")
+	if err != nil {
+		os.RemoveAll(src)
+		t.Fatal(err)
+	}
+	if err := EvalSymlinks(&src); err != nil {
+		os.RemoveAll(tgt)
+		os.RemoveAll(src)
+		t.Fatal(err)
+	}
+	if err := EvalSymlinks(&tgt); err != nil {
+		os.RemoveAll(tgt)
+		os.RemoveAll(src)
+		t.Fatal(err)
+	}
+	defer func() {
+		Unmount(tgt)
+		os.RemoveAll(tgt)
+		os.RemoveAll(src)
+	}()
+	if err := BindMount(src, tgt); err != nil {
+		t.Error(err)
+		t.Fail()
+		return
+	}
+	t.Logf("bind mount success: source=%s, target=%s", src, tgt)
+	mounts, err := GetMounts()
+	if err != nil {
+		t.Error(err)
+		t.Fail()
+		return
+	}
+	success := false
+	for _, m := range mounts {
+		if m.Source == src && m.Path == tgt {
+			success = true
+		}
+		t.Logf("%+v", m)
+	}
+	if !success {
+		t.Errorf("unable to find bind mount: src=%s, tgt=%s", src, tgt)
+		t.Fail()
+	}
+}
+
+func TestGetMounts(t *testing.T) {
+	mounts, err := GetMounts()
+	if err != nil {
+		t.Error(err)
+		t.Fail()
+		return
+	}
+	for _, m := range mounts {
+		t.Logf("%+v", m)
+	}
+}
 
 func TestMountArgs(t *testing.T) {
 	tests := []struct {
